Fixed tests

+++ ckan/bootstrap.sh
@@ -8,7 +8,16 @@ pipenv sync --dev
 ls -la
 cd "$HOME"
 rm -rf venv;
-ln -s .adxvenv/adx_develop-YYZTJvBg venv
-ln -s "$HOME"/venv/bin/ckan /usr/local/bin/ckan
-ln -s "$HOME"/venv/bin/paster /usr/local/bin/ckan-paster
+if [ ! -L venv ]; then
+  ln -s .adxvenv/adx_develop-YYZTJvBg venv
+fi
+if [ ! -L /usr/local/bin/ckan ]; then
+  ln -s "$HOME"/venv/bin/ckan /usr/local/bin/ckan
+fi
+if [ ! -L /usr/local/bin/ckan-paster ]; then
+  ln -s "$HOME"/venv/bin/paster /usr/local/bin/ckan-paster
+fi
+if [ ! -L /usr/local/bin/ckan-pytest ]; then
+  ln -s "$HOME"/venv/bin/pytest /usr/local/bin/ckan-pytest
+fi
 # chown -R ckan:ckan ./
